set for instance only delete

+++ tribble/operations/queue_worker.py
@@ -185,7 +185,8 @@ class MainDisptach(object):
                         state._active()
                     elif cell['job'] in ('schematic_delete',
                                          'zone_delete',
-                                         'redeploy_delete'):
+                                         'redeploy_delete',
+                                         'instance_delete'):
                         if cell.get('zone_id'):
                             state._delete()
                             if cell['job'] == 'redeploy_delete':
@@ -194,12 +195,15 @@ class MainDisptach(object):
                             else:
                                 if 'uuids' in cell and cell['uuids']:
                                     bobs = constructor.MainOffice(nucleus=cell)
+                                    if cell['job'] == 'instance_delete':
+                                        bobs.bob_destroyer()
                                     with STATS.timer('ZoneDelete'):
                                         bobs.bob_destroyer()
                         if cell['job'] == 'schematic_delete':
                             state._delete_resource(skm=True)
                         else:
-                            state._delete_resource()
+                            if not cell['job'] == 'instance_delete':
+                                state._delete_resource()
                     elif cell['job'] == 'reconfig':
                         STATS.incr('Reconfigurations')
                         state._reconfig()
